split the FFI part of rustc_llvm to rustc_llvm::ffi
